A bunch of fiddling around...
Mostly have just been churning through different attempts to understand
how to design and implement supervision trees and ideal ways to register
the processes in such a way as to be able to get "statuses".

In general, this doesn't do anything new that it didn't already and the
code seems to be worse (in terms of readability) BUT it does what it
used to do MUCH faster now, at least, by being more parallelised and
blocking callers less.

About the only NEW thing that I've added here is that the per-PMS syncer
will try to connect to ALL IP addresses and find the first one which is
connectable. This means some servers that seemed inaccessible before are
now accessible.